fix(#416): load consolidation suggestions in HomePage server data
- Add buildConsolidationSuggestions to page-data.ts and call it on
  the fetched subscriptions before returning, replacing the hardcoded
  empty array that caused the initial payload to always be empty.
- Remove duplicate getInitialData definition and duplicate HomePage
  export from page.tsx; page now cleanly delegates to page-data.
